backend/requirements.txt
flask
gunicorn
psycopg2-binary
scikit-learn
pandas
numpy
python-dotenv
prometheus_client
flask-session
pytest
pytest-flask
PyJWT
bcrypt
flask-limiter
google-generativeai
requests